+++ astropy/table/column.py
@@ -709,7 +709,16 @@ class Column(BaseColumn):
     def __setattr__(self, item, value):
         if not isinstance(self, MaskedColumn) and item == "mask":
             raise AttributeError("cannot set mask value to a column in non-masked Table")
-        return super(Column, self).__setattr__(item, value)
+        super(Column, self).__setattr__(item, value)
+
+        if item == 'unit' and issubclass(self.dtype.type, np.number):
+            try:
+                converted = self.parent_table._convert_col_for_table(self)
+            except AttributeError:  # Either no parent table or parent table is None
+                pass
+            else:
+                if converted is not self:
+                    self.parent_table.replace_column(self.name, converted)
 
     def _base_repr_(self, html=False):
         # If scalar then just convert to correct numpy type and use numpy repr